+++ coordinator/app/src/main/kotlin/linea/coordinator/app/conflation/ConflationApp.kt
@@ -43,9 +43,11 @@ import linea.coordinator.clients.prover.ProverClientFactory
 import linea.coordinator.config.toJsonRpcRetry
 import linea.coordinator.config.v2.CoordinatorConfig
 import linea.domain.BlobRecord
+import linea.domain.BlockParameter
 import linea.domain.BlocksConflation
 import linea.encoding.BlockRLPEncoder
 import linea.ethapi.EthApiClient
+import linea.ethapi.EthLogsSearcherImpl
 import linea.ftx.ForcedTransactionsApp
 import linea.metrics.LineaMetricsCategory
 import linea.persistence.AggregationsRepository
@@ -149,11 +151,16 @@ class ConflationApp(
       val contractClient = Web3JLineaRollupSmartContractClientReadOnly(
         contractAddress = configs.protocol.l1.contractAddress,
         web3j = l1Web3jClient,
-        ethLogsClient = createEthApiClient(
-          web3jClient = l1Web3jClient,
-          requestRetryConfig = ftxConfig.l1RequestRetries,
+        ethLogsSearcher = EthLogsSearcherImpl(
           vertx = vertx,
+          ethApiClient = createEthApiClient(
+            web3jClient = l1Web3jClient,
+            requestRetryConfig = ftxConfig.l1RequestRetries,
+            vertx = vertx,
+          ),
         ),
+        finalizedStateSearchInitialBlockParameter = configs.protocol.l1.contractDeploymentBlockNumber
+          ?: BlockParameter.Tag.EARLIEST,
       )
       ForcedTransactionsApp.create(
         config = config,
@@ -359,6 +366,7 @@ class ConflationApp(
         log = LogManager.getLogger("clients.l2.eth.conflation"),
       ),
       ethApiClient = l2EthClient,
+      ethLogsSearcher = EthLogsSearcherImpl(vertx = vertx, ethApiClient = l2EthClient),
       contractAddress = configs.protocol.l2.contractAddress,
       smartContractErrors = configs.smartContractErrors,
       smartContractDeploymentBlockNumber = configs.protocol.l2.contractDeploymentBlockNumber?.number,